refactor: request dto 재설정 및 타입파일 위치 변경
