+++ src/requests/interfaces/request.controller.interface.ts
@@ -1,8 +1,8 @@
-import { checkRequestOutputDTO, IRequest, RequestOutputDTO } from '#requests/request.types.js';
+import { BaseRequestOutputDTO, checkRequestOutputDTO, RequestOutputDTO } from '#requests/types/request.dto.js';
 
 export interface IRequestController {
   checkRequest: (driverId: string) => Promise<checkRequestOutputDTO>;
   getRequest: (requestId: string) => Promise<RequestOutputDTO>;
-  postRequest: (driverId: string) => Promise<IRequest>;
-  deleteRequest: (requestId: string) => Promise<IRequest>;
+  postRequest: (driverId: string) => Promise<BaseRequestOutputDTO>;
+  deleteRequest: (requestId: string) => Promise<BaseRequestOutputDTO>;
 }
