+++ src/requests/interfaces/request.repository.interface.ts
@@ -1,8 +1,9 @@
-import { BaseRequestOutputDTO, CreateRequestDTO, IRequest, PatchRequestDTO } from '#requests/request.types.js';
+import { BaseRequestOutputDTO, CreateRequestDTO, PatchRequestDTO } from '#requests/types/request.dto.js';
 
 export interface IRequestRepository {
   findById: (requestId: string) => Promise<BaseRequestOutputDTO>;
-  create: (data: CreateRequestDTO) => Promise<IRequest>;
-  update: (id: string, data: PatchRequestDTO) => void;
-  delete: (id: string) => Promise<IRequest>;
+  findByMoveInfoId: (moveInfoId: string) => Promise<BaseRequestOutputDTO[]>;
+  create: (data: CreateRequestDTO) => Promise<BaseRequestOutputDTO>;
+  update: (id: string, data: PatchRequestDTO) => Promise<BaseRequestOutputDTO>;
+  delete: (id: string) => Promise<BaseRequestOutputDTO>;
 }
