+++ src/requests/interfaces/request.service.interface.ts
@@ -1,8 +1,8 @@
-import { checkRequestOutputDTO, IRequest, RequestOutputDTO } from '#requests/request.types.js';
+import { BaseRequestOutputDTO, checkRequestOutputDTO, RequestOutputDTO } from '#requests/types/request.dto.js';
 
 export interface IRequestService {
   getRequest: (requestId: string) => Promise<RequestOutputDTO>;
   checkRequest: (driverId: string) => Promise<checkRequestOutputDTO>;
-  postRequest: (driverId: string) => Promise<IRequest>;
-  deleteRequest: (requestId: string) => Promise<IRequest>;
+  postRequest: (driverId: string) => Promise<BaseRequestOutputDTO>;
+  deleteRequest: (requestId: string) => Promise<BaseRequestOutputDTO>;
 }
